requirements.txt
pandas
gspread
google-api-python-client
google-auth
matplotlib
serpapi
python-dotenv
PyYAML
